backend/requirements.txt
black==23.9.1
Django==4.2
django-behaviors==0.5.1
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
drf-spectacular==0.26.4
flake8==6.1.0
flake8-broken-line==1.0.0
flake8-commas==2.1.0
isort==5.12.0
mixer==7.2.2
psycopg2-binary==2.9.7
python-dotenv==1.0.0